+++ src/main/java/io/github/lycoriscafe/yggdrasil/rest/classroom/ClassroomService.java
@@ -16,7 +16,8 @@
 
 package io.github.lycoriscafe.yggdrasil.rest.classroom;
 
-import io.github.lycoriscafe.yggdrasil.commons.*;
+import io.github.lycoriscafe.yggdrasil.commons.CommonService;
+import io.github.lycoriscafe.yggdrasil.commons.Response;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -41,8 +42,8 @@ public class ClassroomService implements EntityService<Classroom> {
                     classrooms.add(new Classroom(
                             resultSet.getInt("grade"),
                             resultSet.getString("name")
-                    ).setId(resultSet.getBigDecimal("id"))
-                            .setTeacherId(resultSet.getBigDecimal("teacherId")));
+                    ).setId(resultSet.getBigInteger("id"))
+                            .setTeacherId(resultSet.getBigInteger("teacherId")));
                 }
             }
