+++ src/main/java/io/github/lycoriscafe/yggdrasil/rest/guardian/Guardian.java
@@ -19,12 +19,12 @@ package io.github.lycoriscafe.yggdrasil.rest.guardian;
 import io.github.lycoriscafe.yggdrasil.commons.Entity;
 import io.github.lycoriscafe.yggdrasil.rest.Gender;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.Objects;
 
 public class Guardian implements Entity {
-    private BigDecimal id;
+    private BigInteger id;
     private String nic;
     private String initName;
     private String fullName;
@@ -50,11 +50,11 @@ public class Guardian implements Entity {
         this.contactNo = Objects.requireNonNull(contactNo);
     }
 
-    public BigDecimal getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public Guardian setId(BigDecimal id) {
+    public Guardian setId(BigInteger id) {
         this.id = id;
         return this;
     }
